product-service/requirements.txt
requests
asgiref==3.8.1
black==25.1.0
click==8.2.1
Django==5.2.1
djangorestframework==3.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
sqlparse==0.5.3
gunicorn
